refactor: remove duplicated interfaces.
