+++ src/App.tsx
@@ -5,8 +5,8 @@ import GenreList from './components/GenreList'
 import { Genre } from './hooks/useGenres'
 import { useState } from 'react'
 import PlatformSelector from './components/PlatformSelector'
-import { Platform } from './hooks/useGames'
 import SortSelector from './components/SortSelector'
+import { Platform } from './hooks/usePlatform'
 import GameHeading from './components/GameHeading'
 import Footer from './components/Footer'
