+++ src/components/PlatformIconList.tsx
@@ -3,8 +3,8 @@ import { MdPhoneIphone } from "react-icons/md";
 import { SiNintendo } from "react-icons/si";
 
 import { HStack, Icon } from "@chakra-ui/react"
-import { Platform } from "../hooks/useGames"
 import { IconType } from "react-icons";
+import { Platform } from "../hooks/usePlatform";
 
 // interface Props {
 //   platforms: Platform[]
