+++ src/components/PlatformSelector.tsx
@@ -1,7 +1,6 @@
 import { Button, Menu, MenuButton, MenuItem, MenuList } from '@chakra-ui/react';
 import { FaChevronDown } from 'react-icons/fa';
-import usePlatform from '../hooks/usePlatform';
-import { Platform } from '../hooks/useGames';
+import usePlatform, { Platform } from '../hooks/usePlatform';
 
 interface Props {
   onSelectPlatform: (platform: Platform) => void;
